pocopine-auth: split monolithic lib.rs into focused modules
The 950-line crate root was hard to navigate once `Role`, `Permission`,
`AuthUser`, `Principal`, `Session`, `RequestContext`, predicate
combinators, async guards, and all their tests piled up in one file.
Split into six focused modules with co-located tests:

- role.rs       Role + Permission
- user.rs       AuthUser
- principal.rs  Principal + Session
- context.rs    RequestContext + ensure_*/require_* guards (host-only)
- provider.rs   AuthError, AuthFuture, AuthProvider, SessionStore
- predicate.rs  Decision, Predicate, combinators, From<Decision>

lib.rs is now a 49-line module roster + `pub use` re-exports. Public
API is byte-for-byte identical; downstream consumers (pocopine-auth-
jwt, pocopine-auth-credentials, pocopine-auth-client, pocopine-server,
pocopine-sync, pocopine-live) build unchanged. Also tightens the
`RequestContext::cookie` rustdoc to point at the `cookie` crate for
RFC 6265 quoted-value needs.
